rename cciptypes to ccip and add new reader iface
