+++ pkg/types/ccip/models.go
@@ -1,4 +1,4 @@
-package cciptypes
+package ccip
 
 import (
 	"encoding/hex"
