+++ pkg/types/ccip/offramp.go
@@ -1,4 +1,4 @@
-package cciptypes
+package ccip
 
 import (
 	"context"
@@ -19,8 +19,6 @@ type OffRampReader interface {
 
 	// GetExecutionStateChangesBetweenSeqNums returns all the execution state change events for the provided message sequence numbers (inclusive).
 	GetExecutionStateChangesBetweenSeqNums(ctx context.Context, seqNumMin, seqNumMax uint64, confirmations int) ([]ExecutionStateChangedWithTxMeta, error)
-
-	GetTokenPoolsRateLimits(ctx context.Context, poolAddresses []Address) ([]TokenBucketRateLimit, error)
 
 	Address() Address
