+++ pkg/types/ccip/onramp.go
@@ -1,4 +1,4 @@
-package cciptypes
+package ccip
 
 import (
 	"context"
