+++ pkg/types/ccip/tokenpool.go
@@ -1,0 +1,7 @@
+package ccip
+
+import "context"
+
+type TokenPoolBatchedReader interface {
+	GetInboundTokenPoolRateLimits(ctx context.Context, tokenPoolReaders []Address) ([]TokenBucketRateLimit, error)
+}
