+++ pkg/types/provider_ccip.go
@@ -3,27 +3,28 @@ package types
 import (
 	"context"
 
-	"github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/ccip/cciptypes"
+	"github.com/smartcontractkit/chainlink-common/pkg/types/ccip"
 )
 
 type CCIPCommitProvider interface {
 	PluginProvider
 
-	NewOnRampReader(ctx context.Context, addr cciptypes.Address) (cciptypes.OnRampReader, error)
-	NewOffRampReader(ctx context.Context, addr cciptypes.Address) (cciptypes.OffRampReader, error)
-	NewCommitStoreReader(ctx context.Context, addr cciptypes.Address) (cciptypes.CommitStoreReader, error)
-	NewPriceRegistryReader(ctx context.Context, addr cciptypes.Address) (cciptypes.PriceRegistryReader, error)
-	NewPriceGetter(ctx context.Context) (cciptypes.PriceGetter, error)
-	SourceNativeToken(ctx context.Context) (cciptypes.Address, error)
+	NewOnRampReader(ctx context.Context, addr ccip.Address) (ccip.OnRampReader, error)
+	NewOffRampReader(ctx context.Context, addr ccip.Address) (ccip.OffRampReader, error)
+	NewCommitStoreReader(ctx context.Context, addr ccip.Address) (ccip.CommitStoreReader, error)
+	NewPriceRegistryReader(ctx context.Context, addr ccip.Address) (ccip.PriceRegistryReader, error)
+	NewPriceGetter(ctx context.Context) (ccip.PriceGetter, error)
+	SourceNativeToken(ctx context.Context) (ccip.Address, error)
 }
 
 type CCIPExecProvider interface {
 	PluginProvider
 
-	NewOnRampReader(ctx context.Context, addr cciptypes.Address) (cciptypes.OnRampReader, error)
-	NewOffRampReader(ctx context.Context, addr cciptypes.Address) (cciptypes.OffRampReader, error)
-	NewCommitStoreReader(ctx context.Context, addr cciptypes.Address) (cciptypes.CommitStoreReader, error)
-	NewPriceRegistryReader(ctx context.Context, addr cciptypes.Address) (cciptypes.PriceRegistryReader, error)
-	NewTokenDataReader(ctx context.Context, tokenAddress cciptypes.Address) (cciptypes.TokenDataReader, error)
-	SourceNativeToken(ctx context.Context) (cciptypes.Address, error)
+	NewOnRampReader(ctx context.Context, addr ccip.Address) (ccip.OnRampReader, error)
+	NewOffRampReader(ctx context.Context, addr ccip.Address) (ccip.OffRampReader, error)
+	NewCommitStoreReader(ctx context.Context, addr ccip.Address) (ccip.CommitStoreReader, error)
+	NewPriceRegistryReader(ctx context.Context, addr ccip.Address) (ccip.PriceRegistryReader, error)
+	NewTokenDataReader(ctx context.Context, tokenAddress ccip.Address) (ccip.TokenDataReader, error)
+	SourceNativeToken(ctx context.Context) (ccip.Address, error)
+	NewTokenPoolBatchedReader(ctx context.Context) (ccip.TokenPoolBatchedReader, error)
 }
